Remove HackerInfo.tsx from Git tracking
