+++ src/components/Hackerinfo.tsx
@@ -1,151 +1,348 @@
-/* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { useState, useEffect, useRef } from "react";
-import { toast } from "react-hot-toast";
-import QrScanner from "qr-scanner";
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { useState } from "react";
+import {
+    Flex,
+    Container,
+    Box,
+    Button,
+    Card,
+    CardBody,
+    Center,
+    Grid,
+    Heading,
+    SimpleGrid,
+    Spacer,
+    Switch,
+    Text,
+    Tabs,
+    TabList,
+    TabPanels,
+    Tab,
+    TabPanel,
+    useColorModeValue,
+} from "@chakra-ui/react";
 import axiosInstance from "../axiosInstance";
-import { Button, Text, Flex } from "@chakra-ui/react";
-import { useUser } from "../components/Authentication";
-import { Navigate } from "react-router-dom";
-import OverridePage from "../components/Manual_Override";
-import HackerInfo from "../components/HackerInfo";
+import { toast } from "react-hot-toast";
 
-const usePage = (initialValue = 0) => {
-    const [page, setPage] = useState(initialValue);
-    const changePage = (pageNumber: number) => setPage(pageNumber);
+type MealId = string;
 
-    return { page, changePage };
-};
+// type Food = Record<MealTime, MealStatus>;
 
-export default function Scanner() {
-    const duplicates = new Set();
-    const [info, setInfo] = useState<any>(null);
-    const [allFood, setAllFood] = useState<any>(null);
-    const [count, setCount] = useState(0);
-    const videoRef = useRef<HTMLVideoElement | null>(null);
-    const { isAuthenticated } = useUser();
-    const { page, changePage } = usePage();
+interface HackerInfoProps {
+    info: any;
+    changePage: any;
+    food: Food;
+}
 
-    const handleScan = async (result: any) => {
-        if (result && result.data != "") {
-            // dedup logic
-            if (duplicates.has(result.data)) return;
-            duplicates.add(result.data);
-            const DEDUP_TIMEOUT_MS = 4000;
-            setTimeout(() => duplicates.delete(result.data), DEDUP_TIMEOUT_MS);
+interface FoodItem {
+    id: string;
+    name: string;
+    day: number;
+    endTime: string;
+}
 
-            // admit
-            const toastId = toast.loading("Admitting...");
-            try {
-                const response = await axiosInstance.post(
-                    "/api/admin/qr/scan",
-                    {
-                        id: result.data,
-                    }
-                );
+interface Food {
+    allFood: Array<FoodItem>;
+    currentMeal: string;
+}
 
-                const foodResponse = await axiosInstance.get("/api/admin/food");
-                const data = response.data;
-                setInfo(data.body);
-                setCount(data.scannedCount);
-                toast.success(data.message, { id: toastId });
+export default function Hackerinfo({
+    info,
+    changePage,
+    food,
+}: HackerInfoProps) {
+    const colors = useColorModeValue(
+        ["#dae1eb", "#dae1eb", "#dae1eb"], // Light mode colors for each tab
+        ["#646973", "#646973", "#646973"] // Dark mode colors
+    );
+    const textColor = useColorModeValue("black", "white"); // For light mode, text is black; for dark mode, text is white
+    const [tabIndex, setTabIndex] = useState(0);
+    const bg = colors[tabIndex];
+    const [displayMeals, setDisplayMeals] = useState<Array<MealId>>([]);
+    const spacing = 6;
 
-                setAllFood(foodResponse.data);
-                console.log(foodResponse.data);
-                console.log(data);
-            } catch (error: any) {
-                toast.error(error.data.fallbackMessage, { id: toastId });
-            }
+    const handleSwitchChange = (mealId: MealId) => {
+        // This method adds food items to displayMeal array when switch is turned on (aka when user eats a meal, this meal is added to displayMeal)
+        // All items in this displayMeal array will be send to backend to be removed
+        // check if mealTime exists, if exists, remove from list, if dont exist, add to list
+        const index = displayMeals.indexOf(mealId);
+        if (index != -1) {
+            // index exists, then remove it
+            setDisplayMeals(
+                displayMeals.filter((item) => item !== displayMeals[index])
+            );
+        } else {
+            // index does not exist, add to list
+            setDisplayMeals([...displayMeals, mealId]);
         }
     };
 
-    useEffect(() => {
-        let qrScanner: QrScanner | null = null;
-        if (videoRef.current) {
-            qrScanner = new QrScanner(
-                videoRef.current,
-                (result) => handleScan(result),
+    // This takes all the food and groups them by day (E.g. day1 groups up dinner only, day2 groups up breakfast, lunch, dinner, day3 groups up breakfast only)
+    const groupFoodByDay = () => {
+        const dayForFood: Record<number, FoodItem[]> = {};
+        for (const item of food.allFood) {
+            if (item.day in dayForFood) {
+                dayForFood[item.day].push(item);
+            } else {
+                dayForFood[item.day] = [item];
+            }
+        }
+        return dayForFood;
+    };
+
+    // returns boolean if food is taken; aka mealId is in user's
+    const isFoodTaken = (mealId: MealId): boolean => {
+        for (const item of info.food) {
+            if (item.serving === mealId) {
+                return true;
+            }
+        }
+        return false;
+    };
+
+    const saveHackerInfo = async () => {
+        // Request body for /api/forms/foodtracker/
+        var food = [];
+        for(const item of displayMeals) {
+            food.push({
+                "application": info.id, // hacker id
+                "serving": item, // servingid for each meal had
+
+            })
+        }
+
+        const toastId = toast.loading("Submitting...");
+        try {
+            const response = await axiosInstance.post(
+                "/api/admin/foodtracker",
                 {
-                    onDecodeError: () => {
-                        // console.error(error);
-                    },
-                    highlightScanRegion: true,
-                    highlightCodeOutline: false,
+                    food: food
                 }
             );
-            qrScanner.start();
+            const data = response.data;
+            toast.success(data.message, { id: toastId });
+        } catch (error: any) {
+            toast.error(error.message, { id: toastId });
         }
-
-        return () => {
-            if (qrScanner) {
-                qrScanner.stop();
-                qrScanner.destroy();
-            }
-        };
-    }, [page]);
-
-    useEffect(() => {
-        if (info != null) {
-            changePage(2);
-        } else {
-            changePage(0);
-        }
-    }, [info]);
-
-    useEffect(() => {}, []);
-
-    if (!isAuthenticated && !import.meta.env.DEV) {
-        return <Navigate to="/login" />;
-    }
-
-    if (page == 1) {
-        return <OverridePage changePage={changePage} />;
-    }
-
-    if (page == 2) {
-        return (
-            <HackerInfo info={info} changePage={changePage} food={allFood} />
-        );
-    }
+        changePage(0);
+    };
 
     return (
-        <Flex
-            style={{
-                height: "100svh",
-                flexDirection: "column",
-                alignItems: "center",
-                // marginTop: "16px",
-                marginRight: "16px",
-                marginLeft: "16px",
-                justifyContent: "space-between",
-                overflow: "hidden",
-            }}
-        >
-            <Flex style={{ flexDirection: "column", gap: "8px" }}>
-                <Text textAlign="center">{count} hackers have checked in!</Text>
-                <video
-                    ref={videoRef}
-                    style={{
-                        width: "50wh",
-                        // maxWidth: "500px",
-                        border: "1px solid black",
-                    }}
-                />
-            </Flex>
-            <Button
-                width="100%"
-                marginBottom="16px"
-                onClick={() => changePage(1)}
-            >
-                Haven't signed up?
-            </Button>
-            <Button
-                width="100%"
-                marginBottom="16px"
-                onClick={() => changePage(2)}
-            >
-                hacker page
-            </Button>
+        <Flex minW="100%" minH="100vh" overflowY="auto" maxH="100vh">
+            <Container pb={10} minH="100vh" fontFamily="mono">
+                {info ? (
+                    <div>
+                        <Box py={7} px={2} width="100%">
+                            <Heading
+                                as="h1"
+                                size="xl"
+                                cursor="default"
+                                mb={spacing}
+                                textAlign="center"
+                            >
+                                Hi,{" "}
+                                {info.answers.firstName +
+                                    " " +
+                                    info.answers.lastName}
+                            </Heading>
+
+                            <Card>
+                                <CardBody>
+                                    <Grid
+                                        templateColumns="3fr"
+                                        w="100%"
+                                        mb={spacing}
+                                        gap={spacing}
+                                        fontSize={16}
+                                        // px={1}
+                                    >
+                                        <Box>
+                                            <Text as="b">Email</Text>
+                                            <Box
+                                                borderWidth="2px"
+                                                borderRadius="lg"
+                                                overflow="hidden"
+                                                p="lg"
+                                                textAlign="center"
+                                                px={2}
+                                            >
+                                                {info.answers.email}
+                                            </Box>
+                                        </Box>
+                                        <Box>
+                                            <Text as="b">Phone Number</Text>
+                                            <Box
+                                                borderWidth="2px"
+                                                borderRadius="lg"
+                                                overflow="hidden"
+                                                textAlign="center"
+                                                px={2}
+                                            >
+                                                {info.answers.phoneNumber}
+                                            </Box>
+                                        </Box>
+                                        <Flex>
+                                            <Box>
+                                                <Center>
+                                                    <Text as="b">Status</Text>
+                                                </Center>
+                                                <Box
+                                                    borderWidth="2px"
+                                                    borderRadius="lg"
+                                                    overflow="hidden"
+                                                    textAlign="center"
+                                                    px={2}
+                                                >
+                                                    {info.applicant.status}
+                                                </Box>
+                                            </Box>
+                                            <Spacer/>
+                                            <Box>
+                                                <Center>
+                                                    <Text as="b">Size</Text>
+                                                </Center>
+                                                <Box
+                                                    borderWidth="2px"
+                                                    borderRadius="lg"
+                                                    overflow="hidden"
+                                                    textAlign="center"
+                                                    px={2}
+                                                >
+                                                    {info.answers.tShirtSize}
+                                                </Box>
+                                            </Box>                                    
+                                        </Flex>
+                                    </Grid>
+                                </CardBody>
+                            </Card>
+
+                            <Card my={spacing}>
+                                <CardBody>
+                                    <Center mb={4}>
+                                        <Text fontSize={18} as="b">Dietary Restrictions</Text>
+                                    </Center>
+                                    <Center>
+                                        <Text fontSize={16}>{info.answers.dietaryRestrictions}</Text>
+                                    </Center>
+                                </CardBody>
+                            </Card>
+
+                            <Card>
+                                <CardBody>
+                                    <Box mb={spacing}>
+                                        <Text fontSize="4xl" as="b">
+                                            Meal Schedule
+                                        </Text>
+                                        <br />
+                                        <Text as="i" fontSize={18}>
+                                            Now Serving: {food.currentMeal ?? "Nothing"}
+                                        </Text>
+                                        <br />
+                                    </Box>
+
+                                    <Tabs
+                                        isManual
+                                        isFitted
+                                        variant="enclosed"
+                                        onChange={(index) => setTabIndex(index)}
+                                        width="100%"
+                                        bg={bg}
+                                        mb={spacing}
+                                    >
+                                        <TabList>
+                                            {["Day 1", "Day 2", "Day 3"].map((day, index) => (
+                                                <Tab
+                                                key={index}
+                                                _focus={{boxShadow: "none"}}
+                                                borderWidth="3px"
+                                                color={textColor}
+                                                >{day}</Tab>
+                                            ))}
+                                        </TabList>
+
+                                        <TabPanels>
+                                        {Object.values(groupFoodByDay()).map((foodItems: FoodItem[], index: number) => {
+                                            {/* {Object.entries(groupFoodByDay()).map(
+                                                ([day, foodItems]) => { */}
+                                                    return (
+                                                        <TabPanel
+                                                            display="flex"
+                                                            flexDirection="column"
+                                                            alignItems="center"
+                                                            justifyContent="center"
+                                                            h="240px"
+                                                            key={index}
+                                                        >
+                                                            <Center>
+                                                                <SimpleGrid
+                                                                    columns={1}
+                                                                    w="100%"
+                                                                >
+                                                                    {foodItems.map(
+                                                                        (foodItem: FoodItem) => {
+                                                                            return (
+                                                                                <Flex
+                                                                                    mb={8}
+                                                                                >
+                                                                                    <Text fontSize="lg">
+                                                                                        {
+                                                                                            foodItem.name
+                                                                                        }
+                                                                                    </Text>
+                                                                                    <Spacer/>
+                                                                                    <Switch
+                                                                                        size="lg"
+                                                                                        ml={12}
+                                                                                        isDisabled={isFoodTaken(
+                                                                                            foodItem.id
+                                                                                        )}
+                                                                                        onChange={() =>
+                                                                                            handleSwitchChange(
+                                                                                                foodItem.id
+                                                                                            )
+                                                                                        }
+                                                                                    />
+                                                                                </Flex>
+                                                                            );
+                                                                        }
+                                                                    )}
+                                                                </SimpleGrid>
+                                                            </Center>
+                                                        </TabPanel>
+                                                    );
+                                                }
+                                            )}
+                                        </TabPanels>
+                                    </Tabs>
+                                </CardBody>
+                            </Card>
+
+                            <Center mt={4}>
+                                <Button
+                                    textAlign="center"
+                                    onClick={saveHackerInfo}
+                                    color={textColor}
+                                    position="fixed"
+                                    bottom="5%"
+                                    left="50%"  // Position at the horizontal middle but does not account for button's width
+                                    transform="translateX(-50%)" // shift the button by half of the button's width to left
+                                    zIndex="1000"
+                                    border="2px"
+                                >
+                                    Save
+                                </Button>
+                            </Center>
+
+                            {/* Add these breaks to make room for save button when fully scrolled down */}
+                            <br />
+                            <br />
+                        </Box>
+                    </div>
+                ) : (
+                    <Text fontSize="lg" >Loading...</Text>
+                )}
+            </Container>
         </Flex>
     );
 }
